SoulMate_BE/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.5
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
mysqlclient==2.2.4
pillow==10.4.0
propcache==0.2.0
PyJWT==2.9.0
PyMySQL==1.1.1
redis==5.1.1
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.3
tzdata==2024.2
urllib3==2.2.3
yarl==1.15.0
